Move benchmark in a separated fileMade some improvments in benchmark forfurture use
